requirements.txt
asgiref==3.3.4
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
python-decouple==3.3
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-filter==2.4.0
# uWSGI==2.0.17.1
